inventario/requirements.txt
asgire==3.10.0
beautifulsoup4==4.14.2
brotli==1.2.0
cffi==2.0.0
crispy-bootstrap4==2025.6
cssselect2==0.8.0
Django==5.2.7
django-allauth==65.13.0
django-bootstrap4==25.2
django-crispy-forms==2.4
django-ranged-response==0.2.0
fonttools==4.60.1
pillow==12.0.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
soupsieve==2.8
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
weasyprint==66.0
webencodings==0.5.1
django-filter==23.2